# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Diet Assistant API
  description: |
    RESTful API for AI-powered diet planning and nutrition management system.
    
    ## Features
    - JWT-based authentication
    - Food inventory management
    - Meal tracking and recording
    - AI-powered meal plan generation
    - Nutrition analysis and statistics
    - Conversation with AI assistant
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication.
    Include the token in the Authorization header: `Bearer {token}`
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9090/api/v1
    description: Development server
  - url: https://api.yourdomain.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: "success"
        timestamp:
          type: integer
          format: int64
          example: 1699999999
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
...
    NutritionData:
      type: object
      properties:
        protein:
          type: number
          format: float
          example: 31.0
        carbs:
          type: number
          format: float
          example: 0.0
        fat:
          type: number
          format: float
          example: 3.6
        fiber:
          type: number
          format: float
          example: 0.0
        calories:
          type: number
          format: float
          example: 165.0
...
    MealFood:
      type: object
      properties:
        food_id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Chicken Breast"
        amount:
          type: number
          format: float
          example: 200
        unit:
          type: string
          example: "g"
    Meal:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        user_id:
          type: integer
          format: int64
          example: 1
        meal_date:
          type: string
          format: date
          example: "2024-01-01"
        meal_type:
          type: string
          enum: [breakfast, lunch, dinner, snack]
          example: "breakfast"
        foods:
          type: array
          items:
            $ref: '#/components/schemas/MealFood'
        nutrition:
          $ref: '#/components/schemas/NutritionData'
        notes:
          type: string
          example: "Healthy breakfast"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /meals:
    get:
      tags:
        - Meals
      summary: List meal records
      description: Get paginated list of meal records
      operationId: listMeals
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: meal_type
          in: query
          schema:
            type: string
            enum: [breakfast, lunch, dinner, snack]
      responses:
        '200':
          description: List of meals
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Meal'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
    
    post:
      tags:
        - Meals
      summary: Create meal record
      description: Add a new meal record
      operationId: createMeal
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - meal_date
                - meal_type
                - foods
              properties:
                meal_date:
                  type: string
                  format: date
                meal_type:
                  type: string
                  enum: [breakfast, lunch, dinner, snack]
                foods:
                  type: array
                  items:
                    type: object
                    properties:
                      food_id:
                        type: integer
                      amount:
                        type: number
                notes:
                  type: string
      responses:
        '201':
          description: Meal created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Meal'
...
  /nutrition/daily/{date}:
    get:
      tags:
        - Nutrition
      summary: Get daily nutrition
      description: Get nutrition statistics for a specific date
      operationId: getDailyNutrition
      security:
        - BearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Daily nutrition statistics
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          date:
                            type: string
                            format: date
                          nutrition:
                            $ref: '#/components/schemas/NutritionData'
                          meal_count:
                            type: integer
...